Convert fcm middleware -> notification service
